requirements.txt
discord.py
pillow
aiohttp
flask
python-dotenv